+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 ## mini.pick
 
 - BREAKING: stop trying to parse path for special format ("path:row" and "path:row:col") if supplied inside a table item. This made impossible working with paths containing ":".
+- FEATURE: respect general URI format for paths inside table items.
 - Update `builtin.files()` to use table items when string item might be ambiguous.
 
 ## mini.test